Patches `replace` operation when `match` is not specified
This is not the world's best patch, but I want to reduce the blast-radius of
these changes for now.

Previously, providing the following valid JSON patch to a recipe:
```json
{
    "op": "replace",
    "path": "@output/build/number",
    "value": "0",
}
```
would result in `Recipe.patch()` wiping out the entire block of `build`
sub-keys.

I'm not entirely sure why that was happening, but I believe it had to do with
defaulting the `match` variable to `.*` which caused the current algorithm to
find every line under build and wipe it completely:
```yaml
build:
  0
  0

```

Tested changes against `perseverance-skills`.

After having run into many issues using `percy`, IMO we need to:
- Make `patch` operations "atomic" (i.e. `.patch()` should not render and save
  changes. To put another way, patching should maintain it's in-memory changes
  and the caller should be responsible for writing to disk. This is to save
  disk I/O operations when we start making lots of changes AND allows a program
  to ask a user if they would like the changes that have been made.
- **In general**, make `percy` [RFC 6902](https://datatracker.ietf.org/doc/html/rfc6902/)
  compliant as much as we possibly can.
  - This will reduce confusion from anyone making calls to `percy`
  - Custom concepts like `@object` are not well documented and do not seem
    entirely necessary.
- Add significant testing automation (primarily in the form of unit testing) to
  ensure reliability of the library AND to prevent regressions.
- Reduce/eliminate the use of `deepcopy()` for performance and code clarity.
  Especially as it seems to be used heavily for logging purposes.
- Reduce/eliminate the injection of custom information into the user-provided
  patch object in favor of using more variables. This correlates heavily with
  the previous point.

I can gladly raise these as GitHub issues, but I wanted them to be written down
in some form. I do not likely have the time to implement all of these ideas
before I get a first version of `recipe-bootstrapper` out, but I am concerned
about the long-term stability of `recipe-bootstrapper` and `percy` if these
points are not addressed.